feat(documents): exponer endpoints rest para la carga, descarga y borrado logico de archivos asociados

+++ fiis/src/main/java/com/sgi/fiis/documentacion/application/usecase/DeactivateDocumentUseCase.java
@@ -15,7 +15,7 @@ public class DeactivateDocumentUseCase {
         this.documentRepositoryPort = documentRepositoryPort;
     }
 
-    public void execute(Long documentId, Integer currentUserId, String currentUserRol) {
+    public void execute(Long documentId, Long currentUserId, String currentUserRol) {
 
         // 1. Buscar el documento en la BD
         Document document = documentRepositoryPort.findById(documentId)
